refactor(rust): split auth crate modules
